dashboard/requirements.txt
flask==3.1.0
snowflake-connector-python==3.12.3
cryptography>=42.0.0
python-dotenv==1.1.0
